refactor: extract CLI into dedicated workspace
