+++ scripts/cli/progress-bar.js
@@ -1,4 +1,4 @@
-import { CliUsageError } from "../../src/plugin/cli/cli-errors.js";
+import { CliUsageError } from "../../src/cli/cli-errors.js";
 
 export const DEFAULT_PROGRESS_BAR_WIDTH = 24;
