+++ src/cli/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "prettier-plugin-gml-cli",
+  "type": "module",
+  "config": {
+    "root_path": "."
+  },
+  "scripts": {
+    "test": "node --test tests/*.test.js"
+  },
+  "dependencies": {
+    "prettier": "^3.6.2"
+  },
+  "engines": {
+    "node": "^18.18.0 || ^20.9.0 || >=21.1.0"
+  }
+}
