+++ src/plugin/src/options/option-utils.js
@@ -1,3 +1,5 @@
+import { normalizeStringList } from "../../../shared/string-utils.js";
+
 function coercePositiveIntegerOption(
     value,
     defaultValue,
@@ -16,52 +18,6 @@ function coercePositiveIntegerOption(
     }
 
     return defaultValue;
-}
-
-const DEFAULT_SPLIT_PATTERN = /[\n,]/;
-
-function normalizeStringList(
-    value,
-    {
-        splitPattern = DEFAULT_SPLIT_PATTERN,
-        allowInvalidType = false,
-        errorMessage = "Value must be provided as a string or array of strings."
-    } = {}
-) {
-    if (value == null) {
-        return [];
-    }
-
-    let entries = null;
-    if (Array.isArray(value)) {
-        entries = value;
-    } else if (typeof value === "string") {
-        const pattern = splitPattern ?? DEFAULT_SPLIT_PATTERN;
-        entries = pattern ? value.split(pattern) : [value];
-    } else if (allowInvalidType) {
-        return [];
-    }
-
-    if (!entries) {
-        throw new TypeError(errorMessage);
-    }
-
-    const seen = new Set();
-
-    for (const entry of entries) {
-        if (typeof entry !== "string") {
-            continue;
-        }
-
-        const trimmed = entry.trim();
-        if (trimmed.length === 0) {
-            continue;
-        }
-
-        seen.add(trimmed);
-    }
-
-    return [...seen];
 }
 
 export { coercePositiveIntegerOption, normalizeStringList };
